requirements.txt
selenium>=4.18
webdriver-manager>=4.0
python-dotenv>=1.0